requirements.txt
google-genai
groq
streamlit
pypdf